Move utility functions and remove obsolete variable.
